refactor: 로또 가격을 Price로 포장
