+++ src/main/java/view/InputView.java
@@ -2,6 +2,7 @@ package view;
 
 import domain.LottoNumber;
 import domain.LottoTicket;
+import domain.Price;
 import java.util.Arrays;
 import java.util.Scanner;
 import java.util.stream.Collectors;
@@ -13,9 +14,9 @@ public class InputView {
         scanner = new Scanner(System.in);
     }
 
-    public int inputPrice() {
+    public Price inputPrice() {
         System.out.println("구입 금액을 입력해 주세요.");
-        return Integer.parseInt(scanner.nextLine());
+        return new Price(Integer.parseInt(scanner.nextLine()));
     }
 
     public LottoTicket getWinnerTicket() {
